// web/src/api/generated/model/storeHistoryEntryRequestPrecision.ts
export type StoreHistoryEntryRequestPrecision =
  (typeof StoreHistoryEntryRequestPrecision)[keyof typeof StoreHistoryEntryRequestPrecision];

export const StoreHistoryEntryRequestPrecision = {
  year: "year",
  month: "month",
  day: "day",
} as const;